+++ godot-client/tests/test_before_delete.gd
@@ -22,6 +22,7 @@ var _db: LocalDatabase
 
 class _TestRow:
 	extends _ModuleTableType
+	const PRIMARY_KEY: StringName = &"id"
 	@export var id: int = 0