+++ godot-client/tests/test_btree_cache.gd
@@ -18,6 +18,7 @@ var _db: LocalDatabase
 
 class _Row:
 	extends _ModuleTableType
+	const PRIMARY_KEY: StringName = &"id"
 	@export var id: int = 0
 	@export var group: int = 0