+++ godot-client/tests/test_pkless_delete_order.gd
@@ -292,6 +292,7 @@ class _FlatRow:
 
 class _KeyedRow:
 	extends _ModuleTableType
+	const PRIMARY_KEY: StringName = &"id"
 	@export var id: int = 0